Display "Create extractor" action only in the context of a message (#7131)
* Display create extractor action only in the context of a message

* Remove only from bindings test

* Unify action type definition for bindings text

* Create separate test files for bindings value and field actions
